+++ threedi_models_and_simulations/processing/algorithms_wq.py
@@ -1,4 +1,4 @@
-from typing import Dict
+from typing import List, Tuple
 from qgis.core import (
     QgsProcessingFeatureSource,
     QgsCoordinateReferenceSystem,
@@ -9,13 +9,16 @@ from qgis.core import (
 from qgis.core import (
     QgsProcessing,
     QgsProcessingAlgorithm,
+    QgsProcessingException,
     QgsProcessingParameterField,
     QgsProcessingParameterFeatureSource,
     QgsProcessingParameterNumber,
     QgsProcessingParameterString,
 )
+from shapely import wkt
 
-from threedi_models_and_simulations.processing.label_rain_zones import label_rain_zones
+from threedi_models_and_simulations.processing.label_dwf import label_dwf
+from threedi_models_and_simulations.processing.label_rain_zones import label_rain_zones, ProcessingException
 from threedi_models_and_simulations.settings import SettingsDialog
 
 
@@ -25,20 +28,20 @@ class MockIFace:
         pass
 
 
-def get_name_wkt_dict(
+def get_name_coordinate_pairs(
         features: QgsProcessingFeatureSource,
         source_crs: QgsCoordinateReferenceSystem,
         name_field: str,
         context
-) -> Dict[str, str]:
+) -> List[Tuple[str, List]]:
     """
-    Returns a {name: wkt} dict (wkt = geometry of the feature in Well-Known Text format)
+    Returns a list of (name: coordinates) pairs
     Transforms the input geometry to WGS84
     Converts curve geometry to linear geometry
     Converts single part to multipart
     Adds postfix to name if geometry is multipart
     """
-    result = dict()
+    result = list()
 
     # Define CRS transformation: source to WGS84
     target_crs = QgsCoordinateReferenceSystem("EPSG:4326")
@@ -53,10 +56,13 @@ def get_name_wkt_dict(
         geom = geom.convertToType(QgsWkbTypes.PolygonGeometry, True)
         for i, geom_part in enumerate(geom.parts()):
             geom_part.transform(transform)
-            wkt = geom_part.asWkt()
+            wkt_string = geom_part.asWkt()
+            shapely_polygon = wkt.loads(wkt_string)
+            rings = [list(shapely_polygon.exterior.coords)]
+            rings += [list(ring.coords) for ring in shapely_polygon.interiors]
             if i > 0:
                 name = name + f" {i + 1}"
-            result[name] = wkt
+            result.append((name, rings))
     return result
 
 
@@ -137,7 +143,6 @@ class SimulateWithRainZonesAlgorithm(QgsProcessingAlgorithm):
             )
         )
 
-
     def processAlgorithm(self, parameters, context, feedback):
         """
         Process the algorithm.
@@ -150,7 +155,7 @@ class SimulateWithRainZonesAlgorithm(QgsProcessingAlgorithm):
 
         polygon_feature_layer.crs()
 
-        name_wkt_dict = get_name_wkt_dict(
+        name_coordinates_dict = get_name_coordinate_pairs(
             features=polygon_feature_source,
             source_crs=polygon_feature_layer.crs(),
             name_field=polygon_name_field,
@@ -158,19 +163,98 @@ class SimulateWithRainZonesAlgorithm(QgsProcessingAlgorithm):
         )
 
         settings_dialog = SettingsDialog(MockIFace())
+        settings_dialog.load_settings()
         api_host = settings_dialog.api_url
         _, api_key = settings_dialog.get_3di_auth()
 
-        label_rain_zones(
+        try:
+            label_rain_zones(
+                api_host,
+                api_key,
+                simulation_template_id=simulation_template_id,
+                simulation_name=simulation_name,
+                zones=name_coordinates_dict,
+                feedback=feedback,
+                max_retries=settings_dialog.upload_timeout
+            )
+        except ProcessingException as e:
+            raise QgsProcessingException(str(e)) from e
+
+        return {}
+
+
+class SimulateWithDWFLabellingAlgorithm(QgsProcessingAlgorithm):
+    """
+    Creates a simulation from a template, adds rain zones to the rain event(s) and add the simulation to the queue
+    """
+
+    SIMULATION_TEMPLATE_ID = "SIMULATION_TEMPLATE_ID"
+    SIMULATION_NAME = "SIMULATION_NAME"
+
+    def createInstance(self):
+        return SimulateWithDWFLabellingAlgorithm()
+
+    def name(self):
+        return 'threedi_simulate_with_dwf_labelling'
+
+    def displayName(self):
+        return 'BETA Simulate with DWF labelling'
+
+    def group(self):
+        return "Simulate"
+
+    def groupId(self):
+        return "simulate"
+
+    def shortHelpString(self):
+        return  f"""
+            <p>Creates a simulation from a template, labels the dry weather flow and adds the simulation to the queue</p>
+            <p>The simulation template must contain dry weather flow.</p>
+            <p>The simulation is owned by the same organisation that owns the simulation from which the template was made</p>
+            <p><i>Note: in the future, this functionality will be integrated into the "New simulation" wizard.</i></p>
+            <h3>Parameters</h3>
+            <h4>Simulation template ID</h4>
+            <p>ID of the simulation template you want to use. Use the simulation wizard to create the simulation you want to run, save it as a template, and copy the simulation template ID to use in this processing algorithm.</p>
+            <h4>Simulation name</h4>
+            <p>Name of the simulation</p>
+            """
+
+    def initAlgorithm(self, config=None):
+        """Define input and output parameters."""
+        self.addParameter(
+            QgsProcessingParameterNumber(
+                self.SIMULATION_TEMPLATE_ID,
+                "Simulation template ID",
+                type=QgsProcessingParameterNumber.Integer
+            )
+        )
+
+        self.addParameter(
+            QgsProcessingParameterString(
+                self.SIMULATION_NAME,
+                "Simulation name"
+            )
+        )
+
+    def processAlgorithm(self, parameters, context, feedback):
+        """
+        Process the algorithm.
+        """
+        simulation_template_id = self.parameterAsInt(parameters, self.SIMULATION_TEMPLATE_ID, context)
+        simulation_name = self.parameterAsString(parameters, self.SIMULATION_NAME, context)
+
+        settings_dialog = SettingsDialog(MockIFace())
+        api_host = settings_dialog.api_url
+        _, api_key = settings_dialog.get_3di_auth()
+
+        label_dwf(
             api_host,
             api_key,
             simulation_template_id=simulation_template_id,
             simulation_name=simulation_name,
-            zones=name_wkt_dict,
             feedback=feedback
         )
 
         return {}
-